# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ServiceResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"anyOf": [],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EquipementResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"serialNumber": {
						"type": "string",
						"nullable": true
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"assignedTo": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_EquipementResponse-Array-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/EquipementResponse"
						},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_EquipementResponse_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/EquipementResponse"
							}
						],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ServiceResponse_EquipementResponse-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/EquipementResponse"
							}
						],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EquipementCreateAttributes": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"serialNumber": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"assignedTo": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EquipementUpdateAttributes": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"serialNumber": {
						"type": "string",
						"nullable": true
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"assignedTo": {
						"type": "string",
						"nullable": true
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"AttendanceResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"userId": {
						"type": "string"
					},
					"checkIn": {
						"type": "string",
						"format": "date-time"
					},
					"checkOut": {
						"type": "string",
						"format": "date-time"
					},
					"date": {
						"type": "string"
					},
					"status": {
						"type": "string"
					},
					"note": {
						"type": "string"
					},
					"user": {},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"userId",
					"checkIn",
					"date",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AttendanceListResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"items": {
							"$ref": "#/components/schemas/AttendanceResponse"
						},
						"type": "array",
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AttendanceSingleResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"message": {
						"type": "string"
					},
					"result": {
						"allOf": [
							{
								"$ref": "#/components/schemas/AttendanceResponse"
							}
						],
						"nullable": true
					},
					"statusCode": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"success",
					"message",
					"result",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CheckOutRequest": {
				"properties": {
					"note": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"UpdateAttendanceRequest": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"present",
							"late",
							"left_early",
							"absent"
						]
					},
					"note": {
						"type": "string"
					},
					"checkOut": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"description": "JWT token (Bearer <token>)"
			}
		}
	},
	"info": {
		"title": "Attendance Backend API",
		"version": "1.0.0",
		"description": "API documentation for attendance management system",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "API Support"
		}
	},
	"paths": {
...
		"/api/equipements": {
			"get": {
				"operationId": "GetAllEquipements",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_EquipementResponse-Array-or-null_"
								}
							}
						}
					},
					"404": {
						"description": "Equipements not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_EquipementResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Equipements"
				],
				"security": [
					{
						"jwt": [
							"equipement:read"
						]
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateEquipement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_EquipementResponse-or-null_"
								}
							}
						}
					},
					"409": {
						"description": "Equipement already exists",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_EquipementResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Equipements"
				],
				"security": [
					{
						"jwt": [
							"equipement:create"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EquipementCreateAttributes"
							}
						}
					}
				}
			}
		},
		"/api/equipements/{equipementId}": {
			"get": {
				"operationId": "GetEquipementById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_EquipementResponse-or-null_"
								}
							}
						}
					},
					"404": {
						"description": "Equipement not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_EquipementResponse-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Equipements"
				],
				"security": [
					{
						"jwt": [
							"equipement:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "equipementId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateEquipement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_EquipementResponse-or-null_"
								}
							}
						}
					},
					"404": {
						"description": "Equipement not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_EquipementResponse-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Equipements"
				],
				"security": [
					{
						"jwt": [
							"equipement:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "equipementId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/EquipementUpdateAttributes"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteEquipement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					},
					"404": {
						"description": "Equipement not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_EquipementResponse-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Equipements"
				],
				"security": [
					{
						"jwt": [
							"equipement:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "equipementId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/equipements/{equipementId}/assign": {
			"put": {
				"operationId": "AssignEquipement",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_EquipementResponse-or-null_"
								}
							}
						}
					},
					"404": {
						"description": "Equipement not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ServiceResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Equipements"
				],
				"security": [
					{
						"jwt": [
							"equipement:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "equipementId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"userId": {
										"type": "string"
									}
								},
								"required": [
									"userId"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/api/attendance": {
			"get": {
				"operationId": "GetAllAttendance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": []
			}
		},
		"/api/attendance/{id}": {
			"get": {
				"operationId": "GetAttendanceById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceSingleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAttendance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceSingleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateAttendanceRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAttendance",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceSingleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:delete"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/attendance/user/{userId}": {
			"get": {
				"operationId": "GetAttendanceByUserId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/attendance/date/{date}": {
			"get": {
				"operationId": "GetAttendanceByDate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "date",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/attendance/report/{startDate}/{endDate}": {
			"get": {
				"operationId": "GetAttendanceReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "startDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "endDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/attendance/report/{userId}/{startDate}/{endDate}": {
			"get": {
				"operationId": "GetUserAttendanceReport",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceListResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:read"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "startDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "endDate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/attendance/{id}/checkout": {
			"put": {
				"operationId": "CheckOut",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AttendanceSingleResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Attendance"
				],
				"security": [
					{
						"jwt": [
							"attendance:update"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CheckOutRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}